requirements.txt
anthropic>=0.45.0
pydantic>=2.0
pyyaml>=6.0
